+++ .gitmodules
@@ -3,7 +3,3 @@
 	url = https://github.com/uuazed/numerapi.git
 	branch = master
 
-[submodule "example-scripts"]
-	path = example-scripts
-	url = https://github.com/numerai/example-scripts.git
-	branch = master
